backend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.2
flask==3.0.3
Flask-Cors==4.0.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib-metadata==8.0.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
openai==0.28.0
pydantic==2.8.2
pydantic-core==2.20.1
pypdf2==3.0.1
python-dotenv==1.0.1
regex==2024.5.15
requests==2.32.3
sniffio==1.3.1
tiktoken==0.7.0
tqdm==4.66.4
typing-extensions==4.12.2
urllib3==2.2.2
werkzeug==3.0.3
yarl==1.9.4
zipp==3.19.2
